1032: refactored and cleaned code and created single camel route
