+++ src/integrationTest/resources/application-leaf-integration.yml
@@ -1,7 +1,8 @@
 start-leaf-route: direct:leaf
-leaf-route-names: judicial-role-type,judicial-contract-type,judicial-region-type,judicial-base-location_type
 
-leaf-route:
+routes-to-execute-leaf: judicial-role-type, judicial-contract-type, judicial-region-type, judicial-base-location_type
+
+route:
   judicial-role-type:
     id: judicial-role-type
     file-name: ${role.file.name}
