--- src/main/java/uk/gov/hmcts/reform/juddata/camel/route/initializer/JrdLeafDataLoadStarter.java
@@ -1,28 +1,0 @@
-package uk.gov.hmcts.reform.juddata.camel.route.initializer;
-
-import lombok.extern.slf4j.Slf4j;
-import org.apache.camel.CamelContext;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.juddata.camel.route.LeafTableRoute;
-import uk.gov.hmcts.reform.juddata.camel.service.AuditProcessingService;
-
-@Component
-@Slf4j
-public class JrdLeafDataLoadStarter {
-
-    @Autowired
-    CamelContext camelContext;
-
-    @Autowired
-    LeafTableRoute leafTableRoutes;
-
-    @Autowired
-    AuditProcessingService schedulerAuditProcessingService;
-
-//    @PostConstruct
-//    public void postConstruct() throws Exception {
-//        camelContext.start();
-//        leafTableRoutes.startRoute();
-//    }
-}
